scripts: add incremental building support to target-qemu
While at it, clean the script and add support for building with different
core os versions.

+++ scripts/build-target-qemu.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #
-# Tested on ubuntu 20.n. You need to have qemu-user-static and binfmt-support
+# Tested on ubuntu 20+. You need to have qemu-user-static and binfmt-support
 # installed.
 #
 
@@ -40,7 +40,6 @@ echo "OpenGL disabled"
 OPENGL="--disable-opengl"
 fi
 
-SPICE_VER="${SPICE_VER:-"spice-0.14.3"}"
 if [ -n "$SPICE" ]; then
 echo "Spice enabled"
 SPICE="--enable-spice"
@@ -65,12 +64,32 @@ echo "VIRGL disabled"
 VIRGL="--disable-virglrenderer"
 fi
 
+#
+# For time being still compile static with 20.04. There is a libc bug present
+# in the 22+ that prevents a large binary static linkage and we can only switch
+# to this after this is merged:
+# https://sourceware.org/bugzilla/show_bug.cgi?id=29514
+#
+# That, or we could do a libc PPA. However the signing infrastructure seems
+# broken too on arm64 (verified with Guillem Jover) so even that is struggle
+# I'm not personally interested in at the moment.
+#
+if [ -n "$STATIC" ]; then
+UBUNTU_BASE=http://cdimage.debian.org/mirror/cdimage.ubuntu.com/ubuntu-base/releases/20.04/release/ubuntu-base-20.04.1-base-arm64.tar.gz
+PKGLIST=`cat $BASE_DIR/scripts/package.list.20`
+else
+UBUNTU_BASE=http://cdimage.debian.org/mirror/cdimage.ubuntu.com/ubuntu-base/releases/22.04/release/ubuntu-base-22.04-base-arm64.tar.gz
+#UBUNTU_BASE=https://cdimage.debian.org/mirror/cdimage.ubuntu.com/ubuntu-base/releases/22.10/release/ubuntu-base-22.10-base-arm64.tar.gz
+PKGLIST=`cat $BASE_DIR/scripts/package.list.22`
+fi
+
 MESA_VER="${MESA_VER:-"mesa-20.2.6"}"
+SPICE_VER="${SPICE_VER:-"spice-0.14.3"}"
+UTIL_LINUX_VER=2.37.4
 
 #
 # Note: cross-compilation is also possible, these can be passed through.
 #
-
 unset CC
 unset LD
 unset CXX
@@ -88,7 +107,6 @@ export PATH=$TOOLDIR/bin:$TOOLDIR/usr/bin:/bin:/usr/bin
 export CHROOTDIR=$BASE_DIR/oss/ubuntu
 
 NJOBS=`nproc`
-PKGLIST=`cat $BASE_DIR/scripts/package.list`
 REPO=`which repo`
 
 set -e
@@ -106,29 +124,25 @@ do_unmount()
 
 do_unmount_all()
 {
-	echo -n "Unmount all binding dirs...: "
-	do_unmount $BASE_DIR/oss/ubuntu/build/qemu
+	[ -n "$LEAVE_MOUNTS" ] && echo "leaving bind mounts in place." && exit 0
+
+	echo "Unmount all binding dirs"
+	do_unmount $CHROOTDIR/build/qemu
 	do_unmount $CHROOTDIR/proc
 	do_unmount $CHROOTDIR/dev
-	echo "DONE!"
 }
 
 do_clean()
 {
 	do_unmount_all
 	cd $BASE_DIR/oss/qemu; sudo git clean -xfd || true
-	sudo rm -rf $BASE_DIR/oss/emu
-	sudo rm -rf $BASE_DIR/oss/ubuntu/build/$SPICE_VER
-	sudo rm -rf $BASE_DIR/oss/ubuntu/build/$MESA_VER
-	sudo rm -rf $BASE_DIR/oss/ubuntu/build/libhybris
 }
 
 do_distclean()
 {
 	do_unmount_all
 	cd $BASE_DIR/oss/qemu; sudo git clean -xfd || true
-	sudo rm -rf $BASE_DIR/oss/emu
-	sudo rm -rf $BASE_DIR/oss/ubuntu
+	sudo rm -rf $CHROOTDIR
 }
 
 do_patch()
@@ -145,35 +159,54 @@ do_patch()
 
 do_sysroot()
 {
-	if [ ! -d $BASE_DIR/oss/ubuntu/build ]; then
-		mkdir -p $BASE_DIR/oss/ubuntu/build
-		cd $BASE_DIR/oss/ubuntu
-		wget -c http://cdimage.debian.org/mirror/cdimage.ubuntu.com/ubuntu-base/releases/20.04/release/ubuntu-base-20.04.1-base-arm64.tar.gz
-		tar xf ubuntu-base-20.04.1-base-arm64.tar.gz
-		echo "nameserver 8.8.8.8" > etc/resolv.conf
-		cp $QEMU_USER usr/bin
-		sudo chmod a+rwx tmp
+	mkdir -p $CHROOTDIR/build
+	if [ -e $CHROOTDIR/bin/bash ]; then
+		sudo mount --bind /dev $CHROOTDIR/dev
+		sudo mount -t proc none $CHROOTDIR/proc
+		return;
 	fi
+
+	cd $CHROOTDIR
+	wget -c $UBUNTU_BASE
+	tar xf `basename $UBUNTU_BASE`
 	sudo mount --bind /dev $CHROOTDIR/dev
 	sudo mount -t proc none $CHROOTDIR/proc
+	echo "nameserver 8.8.8.8" > etc/resolv.conf
+	cp $QEMU_USER usr/bin
+	sudo chmod a+rwx tmp
 	DEBIAN_FRONTEND=noninteractive sudo -E chroot $CHROOTDIR apt-get update
 	DEBIAN_FRONTEND=noninteractive sudo -E chroot $CHROOTDIR apt-get -y install $PKGLIST
+#	sudo -E chroot $CHROOTDIR update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-10 10
+#	sudo -E chroot $CHROOTDIR update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-10 10
+	rm `basename $UBUNTU_BASE`
 }
 
 do_spice()
 {
-	cd $BASE_DIR/oss/ubuntu/build
-	if [ ! -d $BASE_DIR/oss/ubuntu/build/$SPICE_VER ]; then
-		sudo rm -rf $SPICE_VER
-		wget https://www.spice-space.org/download/releases/spice-server/$SPICE_VER.tar.bz2
-		tar xf $SPICE_VER.tar.bz2
-	fi
-	sudo -E chroot $CHROOTDIR sh -c "cd /build/$SPICE_VER; ./configure --prefix=/usr $SSTATIC --disable-celt051 ; make -j$NJOBS ; make install"
+	if [ -e $CHROOTDIR/build/$SPICE_VER/server/.libs/libspice-server.a ]; then return; fi
+
+	cd $CHROOTDIR/build
+	sudo rm -rf $SPICE_VER
+	wget https://www.spice-space.org/download/releases/spice-server/$SPICE_VER.tar.bz2
+	tar xf $SPICE_VER.tar.bz2
+	sudo -E chroot $CHROOTDIR sh -c "cd /build/$SPICE_VER; ./configure --prefix=/usr --enable-static --disable-celt051 ; make -j$NJOBS ; make install"
+}
+
+do_util_linux()
+{
+	if [ -e $CHROOTDIR/build/util-linux-$UTIL_LINUX_VER/dmesg ]; then return; fi
+
+	cd $CHROOTDIR/build
+	wget https://mirrors.edge.kernel.org/pub/linux/utils/util-linux/v2.37/util-linux-$UTIL_LINUX_VER.tar.gz
+	tar xf util-linux-$UTIL_LINUX_VER.tar.gz
+	sudo -E chroot $CHROOTDIR sh -c "cd /build/util-linux-$UTIL_LINUX_VER; ./configure --enable-static ; make -j$NJOBS ; make install"
 }
 
 do_mesa()
 {
-	cd $BASE_DIR/oss/ubuntu/build
+	if [ -d $CHROOTDIR/build/$MESA_VER/build ]; then return; fi
+
+	cd $CHROOTDIR/build
 	wget -c https://archive.mesa3d.org/$MESA_VER.tar.xz
 	tar xf $MESA_VER.tar.xz
 	sudo -E chroot $CHROOTDIR sh -c "cd /build/$MESA_VER; meson build --prefix /usr/local $MESAGL -Dopengl=true -Dosmesa=gallium -Dgallium-drivers=swrast,freedreno $SHARED_GLAPI ; cd build; meson install"
@@ -181,51 +214,34 @@ do_mesa()
 
 do_qemu()
 {
-	mkdir -p $BASE_DIR/oss/ubuntu/build/qemu
-	sudo mount --bind $BASE_DIR/oss/qemu $BASE_DIR/oss/ubuntu/build/qemu
-	mkdir -p $BASE_DIR/oss/ubuntu/build/qemu/build
-	cd $BASE_DIR/oss/ubuntu/build
-	if [ -n "$STATIC" ]; then
-		STATIC_QEMU_LIBS="-L/usr/lib/aarch64-linux-gnu -lgmodule-2.0 -lgobject-2.0 -lgio-2.0 -lglib-2.0 -lspice-server -lpixman-1 -lgmodule-2.0 -lgobject-2.0 -lgio-2.0 -lglib-2.0 -lpthread -lpcre -ljpeg -lm -lffi -lz -lssl -lcrypto -ldl -lopus -lm"
-		sudo -E chroot $CHROOTDIR sh -c "cd /build/qemu/build; ../configure --prefix=/usr --target-list=aarch64-softmmu --with-git-submodules=ignore --extra-ldflags='$STATIC_QEMU_LIBS' --enable-kvm --audio-drv-list=oss --disable-alsa --disable-pa $SPICE $OPENGL $SDL $VIRGL $QSTATIC"
-	else
-		sudo -E chroot $CHROOTDIR sh -c "cd /build/qemu/build; ../configure --prefix=/usr --target-list=aarch64-softmmu --with-git-submodules=ignore --enable-kvm --extra-cflags=\"-I/usr/local/include\" --extra-ldflags=\"-L/usr/local/lib/aarch64-linux-gnu -lgbm\" $SPICE $OPENGL $SDL $VIRGL $QSTATIC"
+	#
+	# Build always
+	#
+	mkdir -p $CHROOTDIR/build/qemu
+	sudo mount --bind $BASE_DIR/oss/qemu $CHROOTDIR/build/qemu
+	mkdir -p $CHROOTDIR/build/qemu/build
+	cd $CHROOTDIR/build
+	if [ ! -e qemu/config.status ]; then
+		if [ -n "$STATIC" ]; then
+			STATIC_QEMU_LIBS="-L/usr/lib/aarch64-linux-gnu -lgmodule-2.0 -lgobject-2.0 -lgio-2.0 -lglib-2.0 -lspice-server -lpixman-1 -lgmodule-2.0 -lgobject-2.0 -lgio-2.0 -lglib-2.0 -lpthread -lpcre -ljpeg -lm -lffi -lz -lssl -lcrypto -ldl -lopus -lm"
+			sudo -E chroot $CHROOTDIR sh -c "cd /build/qemu/build; ../configure --prefix=/usr --target-list=aarch64-softmmu --with-git-submodules=ignore --enable-kvm --extra-cflags=\"-fPIC -fno-stack-protector\" --extra-ldflags='$STATIC_QEMU_LIBS' --audio-drv-list=oss --disable-alsa --disable-pa --disable-pie --disable-vnc $SPICE $OPENGL $SDL $VIRGL $QSTATIC"
+		else
+			sudo -E chroot $CHROOTDIR sh -c "cd /build/qemu/build; ../configure --prefix=/usr --target-list=aarch64-softmmu --with-git-submodules=ignore --enable-kvm --extra-cflags=\"-I/usr/local/include\" --extra-ldflags=\"-L/usr/local/lib/aarch64-linux-gnu -lgbm\" $SPICE $OPENGL $SDL $VIRGL $QSTATIC"
+		fi
 	fi
 	sudo -E chroot $CHROOTDIR sh -c "cd /build/qemu/build; make -j$NJOBS; make install"
 }
 
 do_hybris()
 {
-	if [ -d "$ANDROID_BASE" ]; then
-		cd $BASE_DIR/oss/ubuntu/build
-		sudo rm -rf libhybris
-		git clone https://github.com/libhybris/libhybris.git
-		./libhybris/utils/extract-headers.sh $ANDROID_BASE   $BASE_DIR/oss/ubuntu/usr/local/android/headers
-		sudo -E chroot $CHROOTDIR sh -c "cd /build/libhybris/hybris; ./autogen.sh"
-		sudo -E chroot $CHROOTDIR sh -c "cd /build/libhybris/hybris; ./configure --prefix=/usr --enable-arch=arm64 --enable-adreno-quirks --enable-mesa --enable-ubuntu-linker-overrides --enable-wayland --enable-property-cache --with-android-headers=/usr/local/android/headers; make -j$NJOBS; make install"
-	fi
-}
+	if [ ! -d "$ANDROID_BASE" ]; then return; fi
 
-do_android_emulator()
-{
-	if [ -z "$REPO"  ]; then
-		curl https://storage.googleapis.com/git-repo-downloads/repo > $BASE_DIR/buildtools/usr/bin/repo
-		chmod a+rwx $BASE_DIR/buildtools/usr/bin/repo
-	fi
-	if [ ! -d "$BASE_DIR/oss/emu/external/qemu" ]; then
-		rm -rf $BASE_DIR/oss/emu
-		mkdir $BASE_DIR/oss/emu
-		cd $BASE_DIR/oss/emu
-		repo init -u https://android.googlesource.com/platform/manifest -b emu-master-dev --depth=1
-		repo sync -c --no-tags -j4
-	fi
-	cd $BASE_DIR/oss/emu
-	cd $BASE_DIR/oss/emu/external/qemu
-	[ ! -f /lib/ld-linux-aarch64.so.1 ] && sudo ln -s $CHROOTDIR/usr/lib/aarch64-linux-gnu/ld-linux-aarch64.so.1 /lib/ld-linux-aarch64.so.1
-	export LD_LIBRARY_PATH=$CHROOTDIR/usr/lib/aarch64-linux-gnu/
-	EMUCONFIG="--config release"
-	[ -n "$DEBUG" ] && EMUCONFIG="--config debug"
-	python android/build/python/cmake.py $EMUCONFIG --noqtwebengine --noshowprefixforinfo --target linux_aarch64
+	cd $CHROOTDIR/build
+	sudo rm -rf libhybris
+	git clone https://github.com/libhybris/libhybris.git
+	./libhybris/utils/extract-headers.sh $ANDROID_BASE   $BASE_DIR/oss/ubuntu/usr/local/android/headers
+	sudo -E chroot $CHROOTDIR sh -c "cd /build/libhybris/hybris; ./autogen.sh"
+	sudo -E chroot $CHROOTDIR sh -c "cd /build/libhybris/hybris; ./configure --prefix=/usr --enable-arch=arm64 --enable-adreno-quirks --enable-mesa --enable-ubuntu-linker-overrides --enable-wayland --enable-property-cache --with-android-headers=/usr/local/android/headers; make -j$NJOBS; make install"
 }
 
 do_host_cvd_package()
@@ -265,10 +281,10 @@ fi
 trap do_unmount_all SIGHUP SIGINT SIGTERM EXIT
 
 do_sysroot
+do_util_linux
 do_spice
 [ -z "$STATIC" ] && do_mesa
 do_qemu
-[ -n "$ANDROID_EMU" ] && do_android_emulator
 [ -z "$STATIC" ] && do_hybris
 [ -n "$HOSTCVD" ] && do_host_cvd_package
 echo "All ok!"
